add memmap to np_utils
